REJECT_UNDER_SPOT_COUNT added for indexing, and can now read/write MTZ files from HDF5 files. Only tested in a minor fashion so still very experimental. Addresses in part issue #16

+++ source/ScalingManager.cpp
@@ -165,7 +165,7 @@ void ScalingManager::processReflections(void)
             if (!mtzs[i]->reflection(j)->anyAccepted())
                 continue;
             
-            int refl_id = mtzs[i]->reflection(j)->getReflId();
+            int refl_id = (int)mtzs[i]->reflection(j)->getReflId();
             
             if (mtzs[i]->reflection(j)->getResolution() > 1 / 2.0)
             		continue;
